+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.1-SNAPSHOT
 - use Modifier class to decode method modifiers
+- only call .getMethods when necessary
 
 # 1.1.0 - 13th September 2021
 - breaking changes: return false for values that don't explicitly extend an :extend-via-metadata protocol with zero methods